Refactor navigation configuration
Removed `config/links.ts` and updated paths to use new modular navigation structure under `config/navigation`. This change enhances organization and scalability by separating domain-specific and general navigation configurations.

Took 45 minutes
