+++ app/dashboard/layout.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import { generalSidebarNav } from "@/config/links"
+import { generalSidebarNav } from "@/config/navigation/general"
 import { getCurrentUser } from "@/lib/session"
 import Footer from "@/components/layout/footer"
 import TopNavbar from "@/components/layout/topNavbar"
