+++ app/md/layout.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import { generalSidebarNav } from "@/config/links"
+import { generalNavigation } from "@/config/navigation/general"
 import { getCurrentUser } from "@/lib/session"
 import Footer from "@/components/layout/footer"
 import TopNavbar from "@/components/layout/topNavbar"
@@ -26,7 +26,7 @@ export default async function DashboardLayout({
       />
       <div className="container grid flex-1 gap-12 md:grid-cols-[200px_1fr]">
         <aside className="hidden w-[200px] flex-col md:flex">
-          <SidebarNav items={generalSidebarNav.data} />
+          <SidebarNav items={generalNavigation.sidebarNav.data} />
         </aside>
         <main
           className="flex w-full flex-1 flex-col"
