+++ env.d.ts
@@ -13,8 +13,6 @@ interface Env {
   NEXT_PUBLIC_SITE_OG_IMAGE: string
   NEXT_PUBLIC_API_KEY: string
   NEXT_PUBLIC_API_URL: string
-  NEXT_PUBLIC_DEFAULT_HOMEPAGE: string
-  NEXT_PUBLIC_DEFAULT_AFTER_LOGIN_PATH: string
   NEXTAUTH_URL: string
   NEXTAUTH_SECRET: string
   GOOGLE_CLIENT_ID: string
