+++ env.mjs
@@ -21,8 +21,6 @@ export const env = createEnv({
     NEXT_PUBLIC_SITE_AUTHOR: z.string().optional(),
     NEXT_PUBLIC_SITE_KEYWORDS: z.string().optional(),
     NEXT_PUBLIC_SITE_OG_IMAGE: z.string().url().optional(),
-    NEXT_PUBLIC_DEFAULT_HOMEPAGE: z.string().optional(),
-    NEXT_PUBLIC_DEFAULT_AFTER_LOGIN_PATH: z.string().optional(),
     NEXT_PUBLIC_TEST_DOMAIN: z.string().optional(),
   },
   runtimeEnv: {
@@ -41,8 +39,6 @@ export const env = createEnv({
     NEXT_PUBLIC_SITE_OG_IMAGE: process.env.NEXT_PUBLIC_SITE_OG_IMAGE,
     EMAIL_SERVER: process.env.EMAIL_SERVER,
     EMAIL_FROM: process.env.EMAIL_FROM,
-    NEXT_PUBLIC_DEFAULT_HOMEPAGE: process.env.NEXT_PUBLIC_DEFAULT_HOMEPAGE,
-    NEXT_PUBLIC_DEFAULT_AFTER_LOGIN_PATH: process.env.NEXT_PUBLIC_DEFAULT_AFTER_LOGIN_PATH,
     TEST_DOMAIN: process.env.TEST_DOMAIN,
     NEXT_PUBLIC_TEST_DOMAIN: process.env.NEXT_PUBLIC_TEST_DOMAIN,
   },
